Updated:
BodyParts are now split in to 4 categories, each being able to include a
prefix and a material type. There is also an ItemFactory and some tests
which should give a good idea of how it works. All code but EnemyFactory
has been updated to now exclude the Item class, so we can probably get
rid of it.